--- a/.gitignore
+++ b/.gitignore
@@ -26,3 +26,4 @@ ehthumbs.db
 Thumbs.db
 .vscode/
 *.bak
+*.benchmark