--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -1,12 +1,20 @@
 package hw10programoptimization
 
 import (
-	"encoding/json"
+	"bufio"
+	"errors"
 	"fmt"
 	"io"
-	"io/ioutil"
-	"regexp"
 	"strings"
+
+	jsoniter "github.com/json-iterator/go"
+)
+
+var (
+	ErrInvalidEmail  = errors.New("invalid email")
+	ErrMalformedJSON = errors.New("malformed json")
+
+	json = jsoniter.ConfigCompatibleWithStandardLibrary
 )
 
 type User struct {
@@ -22,46 +30,27 @@ type User struct {
 type DomainStat map[string]int
 
 func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
-	u, err := getUsers(r)
-	if err != nil {
-		return nil, fmt.Errorf("get users error: %w", err)
-	}
-	return countDomains(u, domain)
-}
-
-type users [100_000]User
-
-func getUsers(r io.Reader) (result users, err error) {
-	content, err := ioutil.ReadAll(r)
-	if err != nil {
-		return
-	}
-
-	lines := strings.Split(string(content), "\n")
-	for i, line := range lines {
-		var user User
-		if err = json.Unmarshal([]byte(line), &user); err != nil {
-			return
-		}
-		result[i] = user
-	}
-	return
-}
-
-func countDomains(u users, domain string) (DomainStat, error) {
+	domainSuffix := "." + domain
+	scanner := bufio.NewScanner(r)
 	result := make(DomainStat)
+	user := &User{}
 
-	for _, user := range u {
-		matched, err := regexp.Match("\\."+domain, []byte(user.Email))
-		if err != nil {
-			return nil, err
+	for scanner.Scan() {
+		if err := json.Unmarshal(scanner.Bytes(), user); err != nil {
+			return nil, fmt.Errorf("%s: %w", string(scanner.Bytes()), ErrMalformedJSON)
 		}
 
-		if matched {
-			num := result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])]
-			num++
-			result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])] = num
+		email := strings.ToLower(user.Email)
+		if strings.HasSuffix(email, domainSuffix) {
+			idx := strings.Index(email, "@")
+			if idx == -1 {
+				return nil, fmt.Errorf("%s :%w", user.Email, ErrInvalidEmail)
+			}
+			result[email[idx+1:]]++
 		}
+
+		*user = User{}
 	}
+
 	return result, nil
 }