## Домашнее задание №10 «Оптимизация программы»

### Чек-лист студента ([Что это?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0))
- [x] Я перечитал задание после решения. [Зачем перечитывать задание?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0#user-content-%D0%97%D0%B0%D1%87%D0%B5%D0%BC-%D0%BF%D0%B5%D1%80%D0%B5%D1%87%D0%B8%D1%82%D1%8B%D0%B2%D0%B0%D1%82%D1%8C-%D0%B7%D0%B0%D0%B4%D0%B0%D0%BD%D0%B8%D0%B5)
- [x] Я запустил `go mod tidy`.
- [x] Я удалил `.sync` файл. [Зачем его удалять?](https://github.com/OtusGolang/home_work/wiki/%5B%D0%A1%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0%D0%BC%5D-%D0%9F%D1%80%D0%BE%D1%86%D0%B5%D1%81%D1%81-%D1%81%D0%B4%D0%B0%D1%87%D0%B8-%D0%94%D0%97#user-content-%D0%92%D0%B0%D1%80%D0%B8%D0%B0%D0%BD%D1%82-2)
- [x] Я написал бенчмарк и приложил результаты benchstat.

### Критерии оценки
- [x] Пайплайн зелёный и нет попытки «обмануть» систему - 4 балла
- [x] Добавлены юнит-тесты - до 3 баллов
- [x] Понятность и чистота кода - до 3 баллов

#### Зачёт от 7 баллов

#### Результаты бенчмарков
<details>
<summary>До оптимизаций</summary>

```
goos: linux
goarch: amd64
pkg: github.com/pustato/otus_home_work/hw10_program_optimization
cpu: AMD Ryzen 7 5800X 8-Core Processor             
BenchmarkGetDomainStat-16    	      33	 174174179 ns/op	132856394 B/op	 1700189 allocs/op
PASS
ok  	github.com/pustato/otus_home_work/hw10_program_optimization	6.531s
```
</details>

<details>
<summary>После оптимизаций</summary>

```
goos: linux
goarch: amd64
pkg: github.com/pustato/otus_home_work/hw10_program_optimization
cpu: AMD Ryzen 7 5800X 8-Core Processor             
BenchmarkGetDomainStat-16    	  622183	      9602 ns/op	    6259 B/op	     154 allocs/op
PASS
ok  	github.com/pustato/otus_home_work/hw10_program_optimization	6.075s
```
</details>

<details>
<summary>Результат benchstat</summary>

```
name              old time/op    new time/op    delta
GetDomainStat-16     174ms ± 0%       0ms ± 0%   -99.99%

name              old alloc/op   new alloc/op   delta
GetDomainStat-16     133MB ± 0%       0MB ± 0%  -100.00%

name              old allocs/op  new allocs/op  delta
GetDomainStat-16     1.70M ± 0%     0.00M ± 0%   -99.99%
```
</details>